backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
python-multipart
feedparser
beautifulsoup4
langchain
langchain-openai
openai
langchain-community
itsdangerous
sqlalchemy 